requirements.txt
Flask==2.3.3
SQLAlchemy==1.4.52
pymysql==1.1.1
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.6.0
passlib==1.7.4
python-dotenv==1.0.1
pytest==8.3.3
werkzeug==3.0.4
cryptography==43.0.1
Faker==37.5.3
